Extract update button + unmapping state to props
